server/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
django-cors-headers==4.3.1
six==1.16.0
python-dotenv==1.0.0